Refactor pgquery scripts to be a common script
