+++ services/api/src/admin_new_org.py
@@ -18,7 +18,7 @@ def add_organization(org_spec):
   try:
     org_insert_query = "INSERT INTO public.organizations(name) " \
             f"VALUES ('{org_spec['name']}')"
-    pgquery.insert_db(org_insert_query)
+    pgquery.write_db(org_insert_query)
   except sqlalchemy.exc.IntegrityError as e:
     failed_value = e.orig.args[0]['D']
     failed_value = failed_value.replace('(', '"')
