+++ services/pytest.ini
@@ -5,5 +5,6 @@ pythonpath = .
             addons/images/iss_health_check
             addons/images/rsu_ping_fetch
             api/src
+            common
 env =
       CORS_DOMAIN=test.com
